create lib dir, moved totp_test & middleware
