+++ relay/package.json
@@ -16,7 +16,6 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jasmine": "^3.5.0",
-    "pm2": "^4.2.3",
-    "totp-generator": "0.0.7"
+    "pm2": "^4.2.3"
   }
 }
